+++ src/Controllers/ActivationController.php
@@ -16,14 +16,14 @@
 
 namespace GrahamCampbell\Credentials\Controllers;
 
+use Illuminate\View\Factory;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\URL;
-use GrahamCampbell\Binput\Classes\Binput;
-use GrahamCampbell\Viewer\Classes\Viewer;
-use GrahamCampbell\Queuing\Facades\Queuing;
-use GrahamCampbell\Credentials\Classes\Credentials;
+use GrahamCampbell\Binput\Binput;
+use GrahamCampbell\Credentials\Credentials;
 use GrahamCampbell\Credentials\Providers\UserProvider;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
@@ -36,47 +36,22 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
  * @license    https://github.com/GrahamCampbell/Laravel-Credentials/blob/master/LICENSE.md
  * @link       https://github.com/GrahamCampbell/Laravel-Credentials
  */
-class ActivationController extends AbstractController
+class ActivationController extends BaseController
 {
-    /**
-     * The viewer instance.
-     *
-     * @var \GrahamCampbell\Viewer\Classes\Viewer
-     */
-    protected $viewer;
-
-    /**
-     * The binput instance.
-     *
-     * @var \GrahamCampbell\Binput\Classes\Binput
-     */
-    protected $binput;
-
-    /**
-     * The user provider instance.
-     *
-     * @var \GrahamCampbell\Credentials\Providers\UserProvider
-     */
-    protected $userprovider;
-
     /**
      * Create a new instance.
      *
-     * @param  \GrahamCampbell\Credentials\Classes\Credentials  $credentials
-     * @param  \GrahamCampbell\Viewer\Classes\Viewer  $viewer
-     * @param  \GrahamCampbell\Binput\Classes\Binput  $binput
+     * @param  \GrahamCampbell\Credentials\Credentials  $credentials
+     * @param  \GrahamCampbell\Binput\Binput  $binput
      * @param  \GrahamCampbell\Credentials\Providers\UserProvider  $userprovider
+     * @param  \Illuminate\View\Factory  $view
      * @return void
      */
-    public function __construct(Credentials $credentials, Viewer $viewer, Binput $binput, UserProvider $userprovider)
+    public function __construct(Credentials $credentials, Binput $binput, UserProvider $userprovider, Factory $view)
     {
-        $this->viewer = $viewer;
-        $this->binput = $binput;
-        $this->userprovider = $userprovider;
-
         $this->beforeFilter('throttle.resend', array('only' => array('postResend')));
 
-        parent::__construct($credentials);
+        parent::__construct($credentials, $binput, $userprovider, $view);
     }
 
     /**
@@ -123,7 +98,7 @@ class ActivationController extends AbstractController
      */
     public function getResend()
     {
-        return $this->viewer->make(Config::get('graham-campbell/credentials::resend', 'graham-campbell/credentials::account.resend'));
+        return $this->view->make('graham-campbell/credentials::account.resend');
     }
 
     /**
@@ -148,20 +123,16 @@ class ActivationController extends AbstractController
                     ->with('error', 'That user is already activated.');
             }
 
-            $data = array(
-                'view'    => 'graham-campbell/credentials::emails.resend',
+            $mail = array(
                 'url'     => URL::to(Config::get('graham-campbell/core::home', '/')),
                 'link'    => URL::route('account.activate', array('id' => $user->id, 'code' => $user->getActivationCode())),
                 'email'   => $user->getLogin(),
                 'subject' => Config::get('platform.name').' - Activation'
             );
 
-            try {
-                Queuing::pushMail($data);
-            } catch (\Exception $e) {
-                return Redirect::route('account.resend')->withInput()
-                    ->with('error', 'We were unable to resend the email. Please contact support.');
-            }
+            Mail::queue('graham-campbell/credentials::emails.resend', $mail, function($message) use ($mail) {
+                $message->to($mail['email'])->subject($mail['subject']);
+            });
 
             return Redirect::route('account.resend')
                 ->with('success', 'Check your email for your new activation email.');
@@ -169,35 +140,5 @@ class ActivationController extends AbstractController
             return Redirect::route('account.resend')
                 ->with('error', 'That user does not exist.');
         }
-    }
-
-    /**
-     * Return the viewer instance.
-     *
-     * @return \GrahamCampbell\Viewer\Classes\Viewer
-     */
-    public function getViewer()
-    {
-        return $this->viewer;
-    }
-
-    /**
-     * Return the binput instance.
-     *
-     * @return \GrahamCampbell\Binput\Classes\Binput
-     */
-    public function getBinput()
-    {
-        return $this->binput;
-    }
-
-    /**
-     * Return the user provider instance.
-     *
-     * @return \GrahamCampbell\Credentials\Providers\UserProvider
-     */
-    public function getUserProvider()
-    {
-        return $this->userprovider;
     }
 }