+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGE LOG
 ## V0.3 Alpha (Upcoming - July 2014)
 
 * Upgrade to Laravel 4.2
+* Config cleanup
+* Refactoring
+* Added di aliases
+* Added presenters
 
 
 ## V0.2.1 Alpha (24/05/2014)